Fix warnings

+++ source/app/layout/fastinitiallayout.cpp
@@ -144,7 +144,7 @@ void FastInitialLayout::executeReal(bool)
 
             nodeQueue.push(childNodeId);
 
-            auto h = -1.0f + 2.0f * (i - 1.0f) / static_cast<float>(
+            auto h = -1.0f + 2.0f * (static_cast<float>(i) - 1.0f) / static_cast<float>(
                 static_cast<int>(edgeIds.size()) - 1 + edgeCountOffset);
             auto theta = std::acos(h);
             phi = phi + 3.6f / (std::sqrt((static_cast<float>(
